WIP: Change from indices to coordinate notation
A huge commit that changes the way possible moves are found. Before it was done with differences in indices on the chessboard array. Now it uses vector attacks to determine the validity of the move, confirming that each move stays on the board.

This means that code was overhauled throughout the entire program. The first change was defining the new vector class. This is taken from a previous project and adapted for use here. It does everything you would expect from vectors (fortunately) and is relatively bug-free. 

Next, the ChessBoard class now has compatability with either indices, coordinates or vectors. This will be useful when the pieces start interacting with the board seriously. 

As a result of the change, there were also refactoring of the worse methods and classes. Now there are less "is__" methods and more "assert__" methods. Even a few docstrings were updated.

The chess pieces each had a touch up in order to utilise the new vector attacks.