Refactoring of core.py
Added some extra methods in the vector class to handle the +=, -= and
*= operators, fixed the __abs__ method so that now it uses an internal
private method, plus added a __str__ method to allow for easy printing.
Most importantly, a brand new public method was written for the class
called "unitvector" which gets a quasi-unit vector of the current
vector. I can illustrate what this means with an example.

Assume you have a vector going from file b to file d. In the engine this
is read as (0, 2) for going across two files. Obviously this will return
a unit vector of (0, 1). Now assume there is a vector (3, 3). Using
mathematics, the unit vector would actually be (1/sqrt(2), 1/sqrt(2)).
This is a bad result as we only deal with integers, so it actually
returns (1, 1).

The pieces were altered too. Now the pieces all work with vectors
instead of indices. This change is a result of bugs with finding the
allowed moves from a certain position. This change induced the creation
of lots of new methods:
    - _tovector converts an index to a vector;
    - _toindex converts a vector into an index;
    - distancefromselfto now returns a vector instead of an index;
    - possiblemoves has basically replaced isvalidmove and now also
      returns a list of the possible moves;
    - movetoindex moves the piece to an index;
    - movetovector moves the piece to a vector (i.e. rank/file).

Another benefit of the change is now there is less hackish code and
each of the pieces have almost no extra/overwritten methods besides
the pawns. The PawnPiece class is still a work in progress however.