Complete refactor of ChessBoard class
Now the chessboard is assembled using the OOP mentality of compositon.
To help declutter the core.py file it was all moved to a brand new file.

As such there is a lot of new code. Methods have been written to convert
between vectors, coordinates and indices to any other type to help
create backward compatabilty and to reduce the overhead of using a 2D
list. There is also now a method that can count the number of pieces
between a start and end index, which means the engine is getting closer!

Refactored lots of code and added things like assertion checks etc.
There is also some new classes that inherit from ChessBoard that make
a default game or a user-setup game. These will be useful in debugging
later on.