Fixed basic tests on ChessBoard
Due to the refactoring lots of the code broke. This commit addresses
these issues and restores the testing suite into a working state.

The tests_chesspieces.py is still a WIP but can be included in this
commit. core.py had another two lines added and can also be included.
Made a new file tests_vectors.py which will be the testing suite for
the vector class soon.