Fixed testing suites. Refactored code. Small fixes.
This commit isn't very exciting; only small changes here or there.

Added more tests to the testing suites, specifically the vector suite
which was initalised. Refactored some of the code in the core.py file
and chessboard.py.

Developed some limited framework for the GUI and UI of the program.
In saying that however, there is still a lot to go with them.

Updated both of the files in /docs.