Added UI and GUI.
There is now a working GUI and UI for the board. Now more advanced tests
can be run on chessboard and be easily monitored.

In this commit there is also some updated progress on the code that
handles the logic behind legal moves. This is under heavy construction
and is not in a working state. It is at the top of the current TODO
lists and should be in a working state by the end of the week.