Fixed allpossiblemoves method in _ChessBoardPieces
The allpossiblemoves method was a work-in-progress before this commit.
After some work on it, now the method can handle finding the possible
moves for a single piece or two pieces on a basic level. This means it
can't (yet) figure out checks, pins or pawns but it can handle movement
and captures for the backline pieces.

The allpossiblemoves method only allows for finding the moves for a
certain player: functionality was stripped for both sides as it made
handling the output more difficult then needed.

I also added pretty banners to the sections!