Created brand new, integrated testing suite
Now there is a file where advanced tests are held. These tests are for
things like checking how the chessboard behaves under normal conditions
instead of well defined, strict unit tests.

Current progress in the script is tests for the movement of a single
piece, two pieces on the same side and two pieces on different sides.
They are all quite interactive as it is hard to test specific outcomes
for abstract situations.

Eventually this script will have interactive testing suites and non-
interactive testing suites.