Added ability to detect if a piece is pinned.
To get this functionality, there were a few methods added and others
altered.

First, the intial for loop at the start of allpossiblemoves was
replaced with a method _fetchbasicmovesfor with the only paramter
defining which side you want the moves from.

Next _piecesattackingking was refactored to remove the duplicate code.

The major part of this commit was getting the pins functionality in
allpossiblemoves, which is done by removing moves that leave the king
still under attack. As a result, this code will enable checks and
checkmates to be calculated really quickly.