Code refactor & added method to check move on simulated board
The public call 'move' has been made private and is now called in both
'move' and 'capture' methods. The private method has no checks. This
is to help readability and minimise duplicate code.

Now there is a method that simulates a move on a different board and
checks its legality. Now the code is more readable and less prone to
bugs of not setting the board back into its correct position.