[ADD] En Passant functionality
Unfortunately there was too much dependancy to split this commit up.
Now the engine can handle en passant cases on the board for a variety
of situations, as seen in the integrated tests.

Note that the code hasn't been refactored and is absolutely garbage.
Refactoring the code to a less buggy state is a must.