[REFACTOR] Split _ChessBoardPieces into separate components.
The component _ChessBoardPieces was starting to get too long and look
too hairy. Now that component has been split into even more components
to control each part of the move generation and processing. The new
components are:
	- _ChessBoardPiecesCore: This class controls the tests such as if
      if the king is in check and the pieces that are between the start
      and end points.
    - _ChessBoardCastling: Controls all of the checks and moves for the
      castling rule.
    - _ChessBoardEnPassant: Controls all of the move generation for
      the en passant rule.

A part of this refactor was making the methods more separate and then
combining them back together in the public calls in ChessBoardPieces.