[REFACTOR] Introduced closure on GUI and UI classes
Before this refactor, there were a bunch of methods lying around in the
class that are used once to help readability in the 'main' private
method. Ironically, this actually reduced readability. To counter the
issue now the methods that are used once have been placed in the 'main'
private method itself to be called internally.

These refactors don't impact performace massively as they are called
occassionaly and aren't apart of the bottlenecks or repetitive calls.