[REFACTOR] Refactored the pieces code
This is another major commit; the nature of the beast is that none of
it makes sense unless the refactoring is complete. The refactoring is
still not complete but I need to 'save my progress' before I lose it.

In this commit the methods were moved around and closured for each of
the classes that correspond to the pieces. The _PiecesCore class has
lost a few methods as they moved into ChessBoard_MoveGeneration, with
more being moved in future commits.

The castling class has been refactored into the movement generation
class and the en passant class is next to be moved once refactored some
more.