[REFACTOR] Rest of pieces code
This method is now mostly planned out on where I want stuff to be. In
the end the old method "allpossiblemoves" was removed and refactored
into the "fetchpossiblemovesfor" in the chessboard class. There are
also some other minor changes.

This refactoring requires at least one more "session" to get it up to
coding standards before progressing further with the engine. Currently
we have a bit more of a refactoring plan, but it still isn't completely
right.

Note that the refactoring is incomplete still: there are references to
old methods, attributes and still some poorly optimised code. The
testing suites will now all fail, as they rely on specific calls on the
chessboard. This is the next port of call during this major refactor.

-----------------------------------------------------------------------

This set of refactoring commits is aimed at helping reduce the number
of methods that reside in a class at any one time, thus making
operation and usage much more clear for programmers looking at the
source code. There is still a lot to go with the refactoring but over
time, with more refactoring, the code will look a bit cleaner.