[REFACTOR] Fully refactored piece classes to workable state
The pieces, after being moved, have now been fully refactored into a
clean state. By this, I mean that the pieces have lost their "bloated"
methods, unused attributes and had a general facelift.

As a part in the major refactoring, the chess pieces now are to be
quite basic and dumb; they are now treated similar to the way they are
on a physical board. As such, they have no concept of internal position
on the board. They only know what their allowed directions of motion
are.

Lots of methods have been tagged for review instead of being deleted.
This is because the internal workings of the code are of some use once
the refactoring progresses (i.e. they will go elsewhere) but they don't
below in the pieces class anymore.