[REFACTOR] Completed refactor of chessboardrefactor.py
The new script that contains the chessboard, currently named
chessboardrefactor.py and soon to be chessboard.py, has been refactored
into its new working state.

The chessboard now doesn't concern itself with things like legal moves,
checks, checkmates, stalemates, en passant or even castling. It behaves
as the inanimate object that it is in real life: it simply moves pieces
around and stores them on the board in correct positions. The methods
that were kept also received a bit of a facelift and now are less
bloated.