[REFACTOR] Renamed chessboardrefactor.py to chessboard.py
The old name for the file was used while the code was being refactored.
Now that there is enough changed in the code, the old chessboard.py
(which is now no longer apart of the repository) has been replaced with
the chessboardrefactor.py code.

Note that this doesn't mean that the code is anywhere near complete. It
is just now in a state where it should be used in the program.