[REFACTOR] Initalise new testing suites for refactored code
The old tests were written with the old base code in mind. Instead of
making the tests work with the entirely rewritten base, it is simplier
to just rewrite all of the tests too.

This commit sets up the files that will be used to test the code base.
A part of the commit is adding the old tests into the .gitignore so
that they are no longer apart of the repository.