[REFACTOR] Moved the vector class into core.py
The convert function, which is used extensively throughout the project,
relies on being able to import the vectors into the script. However
since the vectors also need the functions in the core script then core
couldn't import vectors.

To fix this problem the vectors have been moved into the core script.