[ADD] Position class to replace convert function
The convert function, although useful, can be trumped in a few
situations with a static generation of all different forms of the
position on the board. This class does that, taking an input of any of
the options index, coordinate or vector and generates the position in
all of the forms. The converted forms can be accessed as an attribute.