[ADD] Test for duplicateboard method
The method makes an exact copy of the board, which is useful in movegen
and analysis. This test makes sure that after copying the board that
the two instances are completely distinct. This is done by adding a
new piece to the board for the original instance and seeing if there is
a change in the copied board.