[FIX] Moved castle notation if-chain to prevent errors
Before, the position was calculated before determing if there was
castling. This meant that the user had to always pass in a position
when it should just be the castling.

The if-chain has been moved to the front so that now you can simply
state it is castling without anything else.