[FIX] Made finding castling moves work
When undergoing the refactoring, the _castlemoves method broke really
badly. The summary of the breakages are:

    - The illegalmoves method couldn't handle a list of moves, only
      expecting a tuple of integers;
    - Parts of the method forgot that cancastleleft and cancastleright
      moved to the ChessBoard class;
    - Having and empty square where a rook or king were supposed to be
      raised an attribute error;
    - The move function in the ChessBoard class doesn't allow for null
      moves (i.e. not moving) but this method caused that to happen and;
    - The _castlemoves method couldn't handle bad colour inputs.

Other fixes are athestetic, such as changing ugly boolean checks into
functions in the methods like isking and isrook functions. The position
of the king and rooks hav been abstracted and named 'kingpos',
'rookleftpos' and 'rookrightpos' to increase readability.