[FIX] Made illegalmoves work properly
Before I was attempting to determine illegal moves by simulating the
board elsewhere and performing on that. The problem with this was that
all of the methods used the internally stored board state. To fix this
problem, the current board state is stored elsewhere temporarily while
the checks are made and then the old state of the board is restored
once completed.