[ADD] En passant setting during move
Once a move is made on the board, it determines if en passant is in
play. If it is, it will set the file that en passant is on for. Note
that this method does NOT remove en passant once it is used: the user
has to reset it at the end of their turn.