[ADD] Pawn promotion code
Before the pawns could reach the end line and wouldn't be promoted.
Now there is some code that handles that.

Note that this code isn't in the move generator itself, which means
that the engine can't actually consider it as a possibility to promote.
This might mean that at a later date it will have to be moved.