[ADD] Equality methods for chessboards
To be used in the search algorithms, now there are methods in place to
check if boards are in equivalent states. That is, are the pieces in
the exact same positions as each other.

There are two magic methods in place: the __eq__ and __ne__ methods.
There cant be any other magic methods for equality as they make no
sense in this context.