[ADD] Node, tree and search algorithm
Unfortunately these commits couldn't be separate as they are so
interdependent and working with recursive data structures is a new
principle for me so I couldn't think of how to design it from scratch.
Instead this process was very organic and thus the large commit.

First, there is now a node class. These nodes hold the move made to
reach the node, the chessboard state and the entire move history.

Second, there is a tree structure class which contains all of the nodes
in the search. Currently it is very dumb but eventually it could be
extended such that it reduces overhead in the search algorithms.

Finally, a brute search algorithm has been designed. It is very slow,
taking more then 2 seconds to generate the moves of two kings on an
empty board. However it does work and can be extended to work faster
and use less overhead.