[NEW] Performance testing for individual methods
Bottlenecks are starting to rear their ugly heads. This commit adds a
new testing suite to the repository that is only concerned with timing.
It isolates certain methods, executes them 10000 times and determines
the time it took to complete.

It was added as a result of the search algorithm appearing slow, when
in fact it was the move generator.