[REFACTOR] Combine search and evaluate into one class
The reason for this is that I want to periodically evaluate the board.
There are two reasons for this: the first is to make sure the engine
doesn't make the move that hopes for a move, like sacrificing a pawn
in hopes the opponent takes with a queen and; secondly so that I can
implement it into the selective search.

The refactored code passes all of the tests created beforehand and the
altered test_engine.py script refects the new input change.