Removed engine.py to be rewritten from scratch
After battling with the engine for the past week and not getting any
good results, I've decided that I need to rewrite the engine script
from scratch.

The old engine.py file was my first attempt at recursion and evaluating
positions. It was pretty messy at times and at the start I was unsure
how they would interface. With this new knowledge it made sense to
simply start again and make it with these considerations in mind.

The old engine file has been moved into a private file on my own
computer and isn't available in the latest commit.