[ADD] Test for defendingmoves parameter on basicmoves
There has been the addition of the defendingmoves parameter to the
basicmoves method contained in the MoveGenerator class. It adds moves
that are "defending" for the colour. That is to say, if a queen was
connected to a rook it would include the rook in its basicmoves return.
This parameter is used in the engine but not the movegenerator.

This test just makes sure that the new parameter works.