[REFACTOR] New class docstring & type converted to property
There is now a new docstring for the BasePiece class which outlines the
methods and attributes that can be accessed.

Also the type method has been converted to a property, which should help
increase readability. This will have introduced bugs in the code however
so the code is now a WIP until this is fixed.