[REFACTOR] Position class to use properties
Before this commit, the Position class used attributes for the user to
get values and protected them in __setattr__ method. This isn't very
clean code so instead now it works with properties to determine the
index/coordinate/vector and stores the position in a protected attribute
as an index.

A result of using properties is that now the position can be altered if
so desired, but it is just as easy to initalise a new instance of the
class.