[ADD] Reintroduced old enpassant code
I might have bitten off a bit more then I can chew. This code works,
albeit in a messy way, and as such it shouldn't be removed just for the
sake of making it clean. I will only go back and refactor it once it
begins to cause me serious headaches; it looking bad doesn't justify
ripping it out and trying to cram it in somewhere else.