[ADD] Random move code for testing
In order to actually "play" against the computer, I've made it so that
the computer picks a legal move at random to play. This will be a
placeholder until I develop a good megaMax search function. This might
be in place a lot longer then I wanted as the engine is causing a lot of
issues.

This project is going onto the backburner unitl I have finished my other
major projects and tutorials. Sorry to everyone who is following its
progress :(